manejo de errores agregados en carts products y users

+++ src/services/carts.services.js
@@ -1,10 +1,12 @@
 import { CARTS_DAO, PRODUCTS_DAO } from "../dao/index.js";
 import { createTicket as createTicketService } from "./tickets.service.js";
 import CartsRepository from "../repositories/carts.repository.js";
-import ProductsRepository from "../repositories/products.repository.js";
+import * as productsService from './products.services.js'
+import { CartNotFound, CartWithoutStock, ProductNotFound } from "../utils/custom-exceptions.js";
+import { Carts } from "../dao/factory.js";
 
-const cartsRepository = new CartsRepository();
-const productsRepository = new ProductsRepository();
+const cartsDAO = new Carts()
+const cartsRepository = new CartsRepository(cartsDAO);
 
 const getCarts = async () => {
   const carts = await cartsRepository.getAll();
@@ -13,6 +15,9 @@ const getCarts = async () => {
 
 const getCartById = async (cid) => {
   const cart = await cartsRepository.getById(cid);
+  if (!cart) {
+    throw new CartNotFound('Cart not found')
+  }
   return cart;
 };
 
@@ -22,13 +27,20 @@ const createCart = async () => {
 };
 
 const updateCart = async (cid, cart) => {
+  const exists = await getCartById(cid)
+  if (!exists) {
+    throw new CartNotFound('Cart not found')
+  }
   const result = await CARTS_DAO.update(cid, cart);
   return result;
 };
 
 const addToCart = async (cid, pid) => {
   const cart = await getCartById(cid);
-  const product = await productsRepository.getById(pid);
+  if (!cart) {
+    throw new CartNotFound('Cart not found')
+  }
+  const product = await productsService.getProductById(pid);
   const productInCart = cart.products.findIndex(
     (p) => p.product._id.toString() === pid.toString()
   );
@@ -45,6 +57,10 @@ const addToCart = async (cid, pid) => {
 
 const updateQuantity = async (cid, pid, quantity) => {
   const cart = await getCartById(cid);
+  if (!cart) {
+    throw new CartNotFound('Cart not found')
+  }
+  await productsService.getProductById(pid);
   const productInCart = cart.products.findIndex(
     (p) => p.product._id.toString() === pid.toString()
   );
@@ -57,11 +73,17 @@ const updateQuantity = async (cid, pid, quantity) => {
 
 const deleteProductInCart = async (cid, pid) => {
   const cart = await getCartById(cid);
+  if (!cart) {
+    throw new CartNotFound('Cart not found')
+  }
+  await productsService.getProductById(pid);
   const productInCart = cart.products.findIndex(
     (p) => p.product._id.toString() === pid.toString()
   );
   if (productInCart !== -1) {
     cart.products.splice(productInCart, 1);
+  } else {
+    throw new ProductNotFound('Product not found')
   }
   const result = await cartsRepository.deleteProductInCart(cart);
   return result;
@@ -69,6 +91,9 @@ const deleteProductInCart = async (cid, pid) => {
 
 const deleteAllProductsInCart = async (cid) => {
   const cart = await getCartById(cid);
+  if (!cart) {
+    throw new CartNotFound('Cart not found')
+  }
   cart.products.splice(0, cart.products.length);
   const result = await cartsRepository.deleteAllProducts(cart);
   return result;
@@ -76,9 +101,15 @@ const deleteAllProductsInCart = async (cid) => {
 
 const purchase = async (cid, user) => {
   const cart = await getCartById(cid);
+  if (!cart) {
+    throw new CartNotFound('Cart not found')
+  }
   const purchaser = user.email;
   let productsWithoutStock = [];
   let amount = 0;
+  if (cart.products.length <= 0) {
+    throw new CartWithoutStock('Cart without stock')
+  }
   cart.products.forEach(async ({ product, quantity }) => {
     if (product.stock >= quantity) {
       amount += quantity * product.price;
