+++ src/services/products.services.js
@@ -1,7 +1,12 @@
-import { PRODUCTS_DAO } from "../dao/index.js";
+import { Products } from "../dao/factory.js";
 import ProductsRepository from "../repositories/products.repository.js";
+import {
+  IncompleteValues,
+  ProductNotFound,
+} from "../utils/custom-exceptions.js";
 
-const productsRepository = new ProductsRepository();
+const productsDAO = new Products();
+const productsRepository = new ProductsRepository(productsDAO);
 
 const getProducts = async ({ limit, page, sort, category, status }) => {
   if (!limit) {
@@ -36,20 +41,51 @@ const getProducts = async ({ limit, page, sort, category, status }) => {
 
 const getProductById = async (id) => {
   const product = await productsRepository.getById(id);
+  if (!product) {
+    throw new ProductNotFound("Product not found");
+  }
   return product;
 };
 
 const createProduct = async (product) => {
+  if (
+    !product.title ||
+    !product.description ||
+    !product.price ||
+    !product.category ||
+    !product.stock ||
+    !product.code
+  ) {
+    throw new IncompleteValues("Incomplete values");
+  }
   const result = await productsRepository.create(product);
   return result;
 };
 
 const updateProduct = async (id, product) => {
+  const exists = await productsRepository.getById(id);
+  if (!exists) {
+    throw new ProductNotFound("Product not found");
+  }
+  if (
+    !product.title ||
+    !product.description ||
+    !product.price ||
+    !product.category ||
+    !product.stock ||
+    !product.code
+  ) {
+    throw new IncompleteValues("Incomplete values");
+  }
   const result = await productsRepository.update(id, product);
   return result;
 };
 
 const deleteProduct = async (id) => {
+  const exists = await productsRepository.getById(id);
+  if (!exists) {
+    throw new ProductNotFound("Product not found");
+  }
   const result = await productsRepository.delete(id);
   return result;
 };
